+++ src/programs/var.f90
@@ -313,12 +313,14 @@ program midas_var
   ! namelist variables
   integer :: numOuterLoopIterations                    ! number of outer loop iterations (default=1)
   integer :: numIterMaxInnerLoop(maxNumOuterLoopIter)  ! number of each inner loop iterations
+  integer :: numHorizWavebandOuterLoop(maxNumOuterLoopIter) ! number of horizontal wavebands to use for each outer loop iterations
   logical :: limitHuInOuterLoop                        ! impose humidity limits on each outer loop iteration
   logical :: computeFinalNlJo                          ! compute final cost function using non-linear H()
   logical :: useTovsUtil                               ! do channel filtering based on UTIL column of the stats_tovs file
   logical :: writeOuterLoopIncrements                  ! write increments from successive outer loop iterations
   NAMELIST /NAMVAR/ numOuterLoopIterations, numIterMaxInnerLoop, limitHuInOuterLoop
   NAMELIST /NAMVAR/ computeFinalNlJo, useTovsUtil, writeOuterLoopIncrements
+  NAMELIST /NAMVAR/ numHorizWavebandOuterLoop
 
   call ver_printNameAndVersion('var','Variational Assimilation')
 
@@ -353,6 +355,7 @@ program midas_var
   numOuterLoopIterations = 1
   limitHuInOuterLoop = .false.
   numIterMaxInnerLoop(:) = 0
+  numHorizWavebandOuterLoop(:) = MPC_missingValue_INT
   computeFinalNlJo = .false.
   useTovsUtil = .false.
   writeOuterLoopIncrements = .false. 
@@ -377,6 +380,10 @@ program midas_var
   if ( numOuterLoopIterations > 1 .and. &
        .not. all(numIterMaxInnerLoop(1:numOuterLoopIterations) > 0) ) then
     call rti_abort('midas-var: some numIterMaxInnerLoop(:) in namelist are negative or zero')
+  end if
+
+  if ( any(numHorizWavebandOuterLoop(1:numOuterLoopIterations) == MPC_missingValue_INT) ) then
+    call msg('midas-var','Total number of wavebands will be used as defined in NAMBEN', mpiAll_opt=.false.)
   end if
 
   obsMpiStrategy = 'LIKESPLITFILES'
@@ -545,7 +552,8 @@ program midas_var
                        controlVectorIncr, numIterMaxInnerLoop(outerLoopIndex), &
                        deallocHessian_opt=deallocHessian, &
                        isMinimizationFinalCall_opt=isMinimizationFinalCall, &
-                       numIterMaxInnerLoopUsed_opt=numIterMaxInnerLoopUsed )
+                       numIterMaxInnerLoopUsed_opt=numIterMaxInnerLoopUsed, &
+                       numHorizWavebandUsed_opt=numHorizWavebandOuterLoop(outerLoopIndex) )
     numInnerLoopIterDone = numInnerLoopIterDone + numIterMaxInnerLoopUsed
     call msg_memUsage('var')
 
@@ -565,7 +573,8 @@ program midas_var
          dataKind_opt=pre_incrReal, allocHeight_opt=.false., allocPressure_opt=.false.)
 
     ! get final increment with mask if it exists
-    call inc_getIncrement( controlVectorIncr, stateVectorIncr, cvm_nvadim )
+    call inc_getIncrement( controlVectorIncr, stateVectorIncr, cvm_nvadim, &
+                           numHorizWavebandUsed_opt=numHorizWavebandOuterLoop(outerLoopIndex) )
     call gio_readMaskFromFile( stateVectorIncr, './analysisgrid' )
     call msg_memUsage('var')
 
@@ -590,7 +599,8 @@ program midas_var
       call gsv_allocate(stateVectorIncrSum, tim_nstepobsinc, hco_anl, vco_anl, &
            datestamp_opt=tim_getDatestamp(), mpi_local_opt=.true., &
            dataKind_opt=pre_incrReal, allocHeight_opt=.false., allocPressure_opt=.false.)
-      call inc_getIncrement( controlVectorIncrSum, stateVectorIncrSum, cvm_nvadim )
+      call inc_getIncrement( controlVectorIncrSum, stateVectorIncrSum, cvm_nvadim, &
+                             numHorizWavebandUsed_opt=numHorizWavebandOuterLoop(outerLoopIndex) )
       call gio_readMaskFromFile( stateVectorIncrSum, './analysisgrid' )
       call msg_memUsage('var')